Update re:circular reference in docs
Dimitri,

I was walking thru the flow and looks good overall. I have not tried it yet line for line in a lab yet, but just reading the docs Line 71 of  section 3.2.1 goes backward to 3.1.2. This just could be because you did not create a "next step" for vlan only and vlan and logical network" next step. The first option "vlan only" links to "vlan and logical network",  but "vlan and logical network" circular links back to installation. Probably just a not complete yet GIT post, but I need to find time to lab this in one cloud. Good Stuff, thanks!